31st Commit - Adding Test Scenario to cater for all methods in the additional Class LogItem
A method in the LogItem Class was not being bypassed. This is because
there was an if condition comprising of two expressions separated by a
logical OR, and given that within previous tests the first clause was
always found to be true, the second clause was never being executed.
Additionally, a getter in the Account Class was removed from being
commented.