40th Commit - Compound Transaction Class- process Method implemented
The process method for the Compound Transaction has been coded. It was
noted that the process tests need to be changed, whereby the previous
test cases for the process method must be catered for by the
processCompoundTransaction method in the TransactionManager Class. As a
result, the skeleton code for the method was implemented. Furthermore,
the new tests for the process method have been coded and were
successful.