45th Commit- preDefine Mehtod implemented and Tests all Passed
During the implementation of the preDefine Method, it was noticed that
an attribute needed to be added to the Transaction Class,i.e. risk,
accompanied by a setter. Furthermore, the test cases for the preDefine
Method were populated with additional custom source and destination
accounts,which were added to the account database, and were made use of
in the preDefine Method.