46th Commit - RiskPresets Class Implemented for different Transaction Types. The preDefine Methods and Tests adapted to this minor change.
So as to allow the client to add other type of transactions a new
class, named RiskPresets has been created. This class allows you to use
the template presets for high as well as low risk transactions and will
also provide the possibility of creating custom transaction. The Custom
Risk Preset tests have been implemented for the Skeleton code within
its class.Due to this change the preDefine method has minor changes
causing the associated tests to differ slightly as well.